Updated views design, final changes

+++ MyGymWorld/Controllers/UserController.cs
@@ -4,6 +4,7 @@
     using Microsoft.AspNetCore.Mvc;
     using MyGymWorld.Core.Contracts;
     using MyGymWorld.Core.Utilities.Contracts;
+    using MyGymWorld.Data.Models;
     using MyGymWorld.Web.ViewModels.Users;
     
     using static MyGymWorld.Common.NotificationMessagesConstants;
@@ -32,10 +33,19 @@
         [HttpGet]
         public async Task<IActionResult> UserProfile()
         {
-            string userId = this.GetUserId();
-
             try
             {
+                string userId = this.GetUserId();
+
+                ApplicationUser user = await this.userService.GetUserByIdAsync(userId);
+
+                if (user.IsDeleted == true)
+                {
+                    this.TempData[ErrorMessage] = "You were deleted by the Admin!";
+
+                    return this.RedirectToAction("Index", "Home", new { area = "" });
+                }
+
                 UserProfileViewModel userProfileViewModel = await this.userService.GetUserToDisplayByIdAsync(userId);
 
                 return this.View(userProfileViewModel);
@@ -51,26 +61,35 @@
         [HttpPost]
         public async Task<IActionResult> UploadProfilePicture(IFormFile profilePicture)
         {
-            if (profilePicture == null)
-            {
-                this.TempData[ErrorMessage] = "Please choose a picture before submitting!!";
-
-                return this.RedirectToAction(nameof(UserProfile));
-            }
-
-            if (!this.cloudinaryService.IsFileValid(profilePicture))
-            {
-                this.TempData[ErrorMessage] = "The allowed types of pictures are jpg, jpeg and png!";
-
-                return this.RedirectToAction(nameof(UserProfile));
-            }
-
-            string userId = this.GetUserId();
-
-            (string profilePictureUri, string publicId) = await this.userService.GetUserProfilePictureUriAndPublicIdAsync(userId);
-
             try
             {
+                string userId = this.GetUserId();
+
+                ApplicationUser user = await this.userService.GetUserByIdAsync(userId);
+
+                if (user.IsDeleted == true)
+                {
+                    this.TempData[ErrorMessage] = "You were deleted by the Admin!";
+
+                    return this.RedirectToAction("Index", "Home", new { area = "" });
+                }
+
+                if (profilePicture == null)
+                {
+                    this.TempData[ErrorMessage] = "Please choose a picture before submitting!!";
+
+                    return this.RedirectToAction(nameof(UserProfile));
+                }
+
+                if (!this.cloudinaryService.IsFileValid(profilePicture))
+                {
+                    this.TempData[ErrorMessage] = "The allowed types of pictures are jpg, jpeg and png!";
+
+                    return this.RedirectToAction(nameof(UserProfile));
+                }
+
+                (string profilePictureUri, string publicId) = await this.userService.GetUserProfilePictureUriAndPublicIdAsync(userId);
+
                 if (publicId != null)
                 {
                     await this.cloudinaryService.DeletePhotoAsync(publicId);
@@ -86,8 +105,6 @@
             catch (Exception)
             {
                 this.TempData[ErrorMessage] = "Something went wrong when uploading picture!";
-
-                return this.RedirectToAction(nameof(UserProfile));
             }
 
             return this.RedirectToAction(nameof(UserProfile));
@@ -95,27 +112,35 @@
 
         public async Task<IActionResult> DeleteProfilePicture()
         {
-            string userId = this.GetUserId();
-
-            (string profilePictureUri, string publicId) = await this.userService.GetUserProfilePictureUriAndPublicIdAsync(userId);
-
-            if (profilePictureUri == null || publicId == null)
-            {
-                this.TempData[ErrorMessage] = "You don't have profile picture to delete!";
-
-                return this.RedirectToAction(nameof(UserProfile));
-            }
-
             try
             {
+                string userId = this.GetUserId();
+
+                ApplicationUser user = await this.userService.GetUserByIdAsync(userId);
+
+                if (user.IsDeleted == true)
+                {
+                    this.TempData[ErrorMessage] = "You were deleted by the Admin!";
+
+                    return this.RedirectToAction("Index", "Home", new { area = "" });
+                }
+
+                (string profilePictureUri, string publicId) = await this.userService.GetUserProfilePictureUriAndPublicIdAsync(userId);
+
+                if (profilePictureUri == null || publicId == null)
+                {
+                    this.TempData[ErrorMessage] = "You don't have profile picture to delete!";
+
+                    return this.RedirectToAction(nameof(UserProfile));
+                }
+
                 await this.cloudinaryService.DeletePhotoAsync(publicId);
+
                 await this.userService.DeleteUserProfilePictureAsync(userId);
             }
             catch (Exception)
             {
                 this.TempData[ErrorMessage] = "Something went wrong when deleting picture!";
-
-                return this.RedirectToAction(nameof(UserProfile));
             }
 
             return this.RedirectToAction(nameof(UserProfile));
@@ -124,115 +149,141 @@
         [HttpGet]
         public async Task<IActionResult> Edit(string id)
         {
-            EditUserInputModel editUserInputModel = await this.userService.GetUserForEditByIdAsync(id);
+            try
+            {
+                ApplicationUser user = await this.userService.GetUserByIdAsync(id);
 
-            editUserInputModel.CountriesSelectList = await this.countryService.GetAllAsSelectListItemsAsync();
-            editUserInputModel.TownsSelectList = await this.townService.GetAllAsSelectListItemsAsync();
+                if (user.IsDeleted == true)
+                {
+                    this.TempData[ErrorMessage] = "You were deleted by the Admin!";
 
-            return this.View(editUserInputModel);
+                    return this.RedirectToAction("Index", "Home", new { area = "" });
+                }
+
+                EditUserInputModel editUserInputModel = await this.userService.GetUserForEditByIdAsync(id);
+
+                editUserInputModel.CountriesSelectList = await this.countryService.GetAllAsSelectListItemsAsync();
+                editUserInputModel.TownsSelectList = await this.townService.GetAllAsSelectListItemsAsync();
+
+                return this.View(editUserInputModel);
+            }
+            catch (Exception)
+            {
+                this.TempData[ErrorMessage] = "Something went wrong!";
+
+                return this.RedirectToAction(nameof(UserProfile));
+            }
         }
 
 
         [HttpPost]
         public async Task<IActionResult> Edit(string id, EditUserInputModel editUserInputModel)
         {
-            editUserInputModel.CountriesSelectList = await this.countryService.GetAllAsSelectListItemsAsync();
-            editUserInputModel.TownsSelectList = await this.townService.GetAllAsSelectListItemsAsync();
-
-            if (!this.ModelState.IsValid)
-            {
-                return this.View(editUserInputModel);
-            }
-
-            if (editUserInputModel.FirstName != null && editUserInputModel.LastName == null)
-            {
-                this.ModelState.AddModelError("LastName", "LastName is required when you have FirstName!");
-
-                return this.View(editUserInputModel);
-            }
-
-            if (editUserInputModel.FirstName == null && editUserInputModel.LastName != null)
-            {
-                this.ModelState.AddModelError("FirstName", "FirstName is required when you have LastName!");
-
-                return this.View(editUserInputModel);
-            }
-
-            if (editUserInputModel.PhoneNumber != null)
-            {
-                bool userExistsByPhoneNumber = await this.userService.CheckIfUserExistsByPhoneNumberAsync(editUserInputModel.PhoneNumber);
-
-                if (userExistsByPhoneNumber)
-                {
-                    this.ModelState.AddModelError("PhoneNumber", "User with this phone exists!");
-                    
-                    return this.View(editUserInputModel);
-                }
-            }
-
-            if (!string.IsNullOrWhiteSpace(editUserInputModel.Address))
-            {
-                if (editUserInputModel.CountryId == "None")
-                {
-                    this.ModelState.AddModelError("CountryId", "Country is required when you have address!");
-                   
-                    return this.View(editUserInputModel);
-                }
-
-                if (editUserInputModel.TownId == "None")
-                {
-                    this.ModelState.AddModelError("TownId", "Town is required when you have address!");
-
-                    return this.View(editUserInputModel);
-                }
-
-                bool isPresent = await this.townService.CheckIfTownIsPresentByCountryIdAsync(editUserInputModel.TownId!, editUserInputModel.CountryId!);
-
-                if (isPresent == false)
-                {
-                    this.ModelState.AddModelError("TownId", "The town should be in the chosen country!");
-
-                    return this.View(editUserInputModel);
-                }
-            }
-
-            if (string.IsNullOrWhiteSpace(editUserInputModel.Address))
-            {
-                if (editUserInputModel.CountryId != "None"
-                    && editUserInputModel.TownId != null)
-                {
-                    this.ModelState.AddModelError("CountryId", "You cannot choose a country without an address!");
-                    this.ModelState.AddModelError("TownId", "You cannot choose a town without an address!");
-                    
-                    return this.View(editUserInputModel);
-                }
-                else if (editUserInputModel.CountryId != "None")
-                {
-                    this.ModelState.AddModelError("CountryId", "You cannot choose a country without an address!");
-
-                    return this.View(editUserInputModel);
-                }
-                else if (editUserInputModel.TownId != null 
-                    && editUserInputModel.TownId != "None")
-                {
-                    this.ModelState.AddModelError("TownId", "You cannot choose a town without an address!");
-
-                    return this.View(editUserInputModel);
-                }
-            }
-
-            bool isUsernamePresent = await this.userService.CheckIfUserExistsByUsernameAsync(editUserInputModel.UserName);
-            var user = await this.userService.GetUserByIdAsync(id);
-
-            if (isUsernamePresent && user.Id != Guid.Parse(id))
-            {
-                this.ModelState.AddModelError("UserName", "This username is already taken!");
-
-                return this.View(editUserInputModel);
-            }
-
             try
             {
+                ApplicationUser user = await this.userService.GetUserByIdAsync(id);
+
+                if (user.IsDeleted == true)
+                {
+                    this.TempData[ErrorMessage] = "You were deleted by the Admin!";
+
+                    return this.RedirectToAction("Index", "Home", new { area = "" });
+                }
+
+                editUserInputModel.CountriesSelectList = await this.countryService.GetAllAsSelectListItemsAsync();
+                editUserInputModel.TownsSelectList = await this.townService.GetAllAsSelectListItemsAsync();
+
+                if (!this.ModelState.IsValid)
+                {
+                    return this.View(editUserInputModel);
+                }
+
+                if (editUserInputModel.FirstName != null && editUserInputModel.LastName == null)
+                {
+                    this.ModelState.AddModelError("LastName", "LastName is required when you have FirstName!");
+
+                    return this.View(editUserInputModel);
+                }
+
+                if (editUserInputModel.FirstName == null && editUserInputModel.LastName != null)
+                {
+                    this.ModelState.AddModelError("FirstName", "FirstName is required when you have LastName!");
+
+                    return this.View(editUserInputModel);
+                }
+
+                if (editUserInputModel.PhoneNumber != null)
+                {
+                    bool userExistsByPhoneNumber = await this.userService.CheckIfUserExistsByPhoneNumberAsync(editUserInputModel.PhoneNumber);
+
+                    if (userExistsByPhoneNumber)
+                    {
+                        this.ModelState.AddModelError("PhoneNumber", "User with this phone exists!");
+
+                        return this.View(editUserInputModel);
+                    }
+                }
+
+                if (!string.IsNullOrWhiteSpace(editUserInputModel.Address))
+                {
+                    if (editUserInputModel.CountryId == "None")
+                    {
+                        this.ModelState.AddModelError("CountryId", "Country is required when you have address!");
+
+                        return this.View(editUserInputModel);
+                    }
+
+                    if (editUserInputModel.TownId == "None")
+                    {
+                        this.ModelState.AddModelError("TownId", "Town is required when you have address!");
+
+                        return this.View(editUserInputModel);
+                    }
+
+                    bool isPresent = await this.townService.CheckIfTownIsPresentByCountryIdAsync(editUserInputModel.TownId!, editUserInputModel.CountryId!);
+
+                    if (isPresent == false)
+                    {
+                        this.ModelState.AddModelError("TownId", "The town should be in the chosen country!");
+
+                        return this.View(editUserInputModel);
+                    }
+                }
+
+                if (string.IsNullOrWhiteSpace(editUserInputModel.Address))
+                {
+                    if (editUserInputModel.CountryId != "None"
+                        && editUserInputModel.TownId != null)
+                    {
+                        this.ModelState.AddModelError("CountryId", "You cannot choose a country without an address!");
+                        this.ModelState.AddModelError("TownId", "You cannot choose a town without an address!");
+
+                        return this.View(editUserInputModel);
+                    }
+                    else if (editUserInputModel.CountryId != "None")
+                    {
+                        this.ModelState.AddModelError("CountryId", "You cannot choose a country without an address!");
+
+                        return this.View(editUserInputModel);
+                    }
+                    else if (editUserInputModel.TownId != null
+                        && editUserInputModel.TownId != "None")
+                    {
+                        this.ModelState.AddModelError("TownId", "You cannot choose a town without an address!");
+
+                        return this.View(editUserInputModel);
+                    }
+                }
+
+                bool isUsernamePresent = await this.userService.CheckIfUserExistsByUsernameAsync(editUserInputModel.UserName);
+
+                if (isUsernamePresent && user.Id.ToString() != id)
+                {
+                    this.ModelState.AddModelError("UserName", "This username is already taken!");
+
+                    return this.View(editUserInputModel);
+                }
+
                 await this.userService.EditUserAsync(editUserInputModel.Id, editUserInputModel);
             }
             catch (Exception ex)
